# api/openapi.yaml
openapi: "3.1.0"
info:
  title: "Kinetic v3 API"
  version: "0.1.0"
  description: |
    REST API for the Kinetic v3 Rocket League esports league management platform.

    ## Authentication
    Most read endpoints are publicly accessible. Write and sensitive operations
    require authentication via either a session cookie (set by the OAuth login flow)
    or an HTTP Bearer token (API token issued via `/v1/api-tokens`).

    ## Permission model
    Protected routes enforce a resource/action pair checked against the caller's
    roles. Available resources are defined in `authz.resources` and include
    `league`, `franchise`, `club`, `team`, `player`, `roster_membership`,
    `role_assignment`, `queue`, `queue_entry`, `queue_ban`, `scrim`,
    `scrim_promotion`, `result_submission`, `result_override`, `replay_evidence`,
    `platform_account_link`, `player_rating`, `rating_adjustment`,
    `matchmaking_decision`, `exception_ticket`, `season`, `schedule_group`,
    `fixture`, `match`, `eligibility`, `player_stat`, `skill_group`,
    `skill_group_transition`, `organization_config`, `api_token`, and
    `player_notification`.
servers:
  - url: "http://localhost:8080"
    description: "Local development server"
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: kinetic_session
      description: "Signed session token set by the OAuth login flow."
    bearerToken:
      type: http
      scheme: bearer
      description: "Long-lived API token issued via POST /v1/api-tokens."
  schemas:
...
    SessionPrincipal:
      type: object
      properties:
        subject:
          type: string
          description: "Unique identifier for the authenticated user (e.g. discord:<snowflake>)."
        displayName:
          type: string
        roles:
          type: array
          items:
            type: string
      required:
        - subject
        - displayName
        - roles

    # ── Hierarchy ──────────────────────────────────────────────────────────────
...
    RosterMembership:
      type: object
      properties:
        id:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        isActive:
          type: boolean
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - playerId
        - teamId
        - isActive
        - createdAt
...
    QueueEntry:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        teamId:
          type: integer
          format: int64
        isActive:
          type: boolean
        stage:
          type: integer
          format: int32
        createdAt:
          type: string
          format: date-time
        stageAt:
          type: string
          format: date-time
        leftAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - queueId
        - teamId
        - isActive
        - stage
        - createdAt
        - stageAt
    QueueBan:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        bannedByActor:
          type: string
        banReason:
          type: string
        isActive:
          type: boolean
        bannedAt:
          type: string
          format: date-time
        unbannedByActor:
          type: string
          nullable: true
        unbanReason:
          type: string
          nullable: true
        unbannedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - queueId
        - playerId
        - bannedByActor
        - banReason
        - isActive
        - bannedAt

    # ── Scrim ──────────────────────────────────────────────────────────────────
    Scrim:
      type: object
      properties:
        id:
          type: integer
          format: int64
        queueId:
          type: integer
          format: int64
        homeTeamId:
          type: integer
          format: int64
        awayTeamId:
          type: integer
          format: int64
        state:
          type: string
          description: "State machine value (e.g. popped, checked_in, in_progress, complete, cancelled)."
        lobbyName:
          type: string
          nullable: true
        lobbyPassword:
          type: string
          nullable: true
        poppedAt:
          type: string
          format: date-time
          nullable: true
        homeCheckedInAt:
          type: string
          format: date-time
          nullable: true
        awayCheckedInAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
        startedAt:
          type: string
          format: date-time
          nullable: true
        endedAt:
          type: string
          format: date-time
          nullable: true
      required:
        - id
        - queueId
        - homeTeamId
        - awayTeamId
        - state
        - createdAt
...
    ResultSubmission:
      type: object
      properties:
        id:
          type: integer
          format: int64
        contextType:
          type: string
          description: "Either 'scrim' or 'match'."
        contextId:
          type: integer
          format: int64
        submittedByTeamId:
          type: integer
          format: int64
        homeTeamId:
          type: integer
          format: int64
        awayTeamId:
          type: integer
          format: int64
        winningTeamId:
          type: integer
          format: int64
        losingTeamId:
          type: integer
          format: int64
        state:
          type: string
          description: "Workflow state (e.g. pending, ratified, rejected, overridden)."
        payloadJson:
          description: "Arbitrary JSON payload attached to the submission."
        homeRatifiedAt:
          type: string
          format: date-time
          nullable: true
        awayRatifiedAt:
          type: string
          format: date-time
          nullable: true
        rejectedByTeamId:
          type: integer
          format: int64
          nullable: true
        rejectionReason:
          type: string
          nullable: true
        rejectedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - contextType
        - contextId
        - submittedByTeamId
        - homeTeamId
        - awayTeamId
        - winningTeamId
        - losingTeamId
        - state
        - createdAt
...
    EligibilityProjectionPoint:
      type: object
      properties:
        effectiveAt:
          type: string
          format: date-time
        points:
          type: integer
          format: int32
        isEligible:
          type: boolean
      required:
        - effectiveAt
        - points
        - isEligible
    EligibilityStatus:
      type: object
      properties:
        subject:
          type: string
        points:
          type: integer
          format: int32
        thresholdPoints:
          type: integer
          format: int32
        decayPerWeek:
          type: integer
          format: int32
        eligibleUntil:
          type: string
          format: date-time
        evaluatedAt:
          type: string
          format: date-time
        projection:
          type: array
          items:
            $ref: "#/components/schemas/EligibilityProjectionPoint"
      required:
        - subject
        - points
        - thresholdPoints
        - decayPerWeek
        - eligibleUntil
        - evaluatedAt
        - projection

    # ── Notifications ──────────────────────────────────────────────────────────
    PlayerNotification:
      type: object
      properties:
        id:
          type: integer
          format: int64
        playerId:
          type: integer
          format: int64
        category:
          type: string
          description: >
            One of: scrim_popped, queue_ban, ratification_needed,
            skill_group_change, roster_change.
        contextType:
          type: string
          nullable: true
        contextId:
          type: integer
          format: int64
          nullable: true
        message:
          type: string
        readAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - playerId
        - category
        - message
        - createdAt

    # ── API Tokens ─────────────────────────────────────────────────────────────
    APIToken:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        tokenPrefix:
          type: string
          description: "First few characters of the token, shown for identification only."
        subject:
          type: string
        scopes:
          type: array
          items:
            type: string
        createdBy:
          type: integer
          format: int64
          nullable: true
        expiresAt:
          type: string
          format: date-time
          nullable: true
        revokedAt:
          type: string
          format: date-time
          nullable: true
        lastUsedAt:
          type: string
          format: date-time
          nullable: true
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - tokenPrefix
        - subject
        - scopes
        - createdAt
    APITokenCreateResponse:
      allOf:
        - $ref: "#/components/schemas/APIToken"
        - type: object
          properties:
            plainTextToken:
              type: string
              description: "The raw token value. Only returned once at creation time."
          required:
            - plainTextToken

    # ── Provider info ──────────────────────────────────────────────────────────
...
paths:
...
  /v1/me:
    get:
      operationId: getMe
      summary: "Get the current session principal"
      tags:
        - me
      security:
        - sessionCookie: []
      responses:
        "200":
          description: "Current principal."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SessionPrincipal"
        "401":
          description: "Not authenticated."
  /v1/me/eligibility:
    get:
      operationId: getMeEligibility
      summary: "Get eligibility status for the current session player"
      tags:
        - me
      security:
        - sessionCookie: []
      responses:
        "200":
          description: "Eligibility status."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EligibilityStatus"
        "401":
          description: "Not authenticated."
        "503":
          description: "Hierarchy store unavailable."
  /v1/me/roster-membership:
    get:
      operationId: getMeRosterMembership
      summary: "Get the active roster membership for a player"
      tags:
        - me
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Active roster membership (null if none)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RosterMembership"
        "400":
          description: "player_id query parameter required."
        "503":
          description: "Hierarchy store unavailable."
  /v1/me/queue-bans:
    get:
      operationId: getMeQueueBans
      summary: "Get active queue bans for a player"
      tags:
        - me
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Active queue bans."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/QueueBan"
        "400":
          description: "player_id query parameter required."
        "503":
          description: "Hierarchy store unavailable."
  /v1/me/queue-entry:
    get:
      operationId: getMeQueueEntry
      summary: "Get the active queue entry for a player"
      tags:
        - me
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Active queue entry (null if not queued)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QueueEntry"
        "400":
          description: "player_id query parameter required."
        "503":
          description: "Hierarchy store unavailable."
  /v1/me/scrim:
    get:
      operationId: getMeScrim
      summary: "Get the active scrim for a player"
      tags:
        - me
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Active scrim (null if none)."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Scrim"
        "400":
          description: "player_id query parameter required."
        "503":
          description: "Hierarchy store unavailable."
  /v1/me/result-submissions:
    get:
      operationId: getMeResultSubmissions
      summary: "List result submissions for the current player's team"
      tags:
        - me
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: state
          in: query
          required: false
          description: "Filter by submission state."
          schema:
            type: string
      responses:
        "200":
          description: "Result submissions."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ResultSubmission"
        "400":
          description: "player_id query parameter required."
        "503":
          description: "Hierarchy store unavailable."

  # ── Notifications ──────────────────────────────────────────────────────────
  /v1/me/notifications:
    get:
      operationId: listMeNotifications
      summary: "List notifications for a player"
      tags:
        - notifications
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
        - name: unread
          in: query
          required: false
          description: "If 'true', return only unread notifications."
          schema:
            type: string
      responses:
        "200":
          description: "Notifications."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PlayerNotification"
        "400":
          description: "player_id query parameter required."
        "503":
          description: "Notifications store unavailable."
  /v1/me/notifications/mark-read:
    post:
      operationId: markNotificationsRead
      summary: "Mark one or more notifications as read"
      tags:
        - notifications
      parameters:
        - name: player_id
          in: query
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                ids:
                  type: array
                  items:
                    type: integer
                    format: int64
                  description: "Specific notification IDs to mark read. Ignored when all is true."
                all:
                  type: boolean
                  description: "If true, mark all notifications for the player as read."
      responses:
        "204":
          description: "Notifications marked as read."
        "400":
          description: "player_id query parameter required or invalid request body."
        "503":
          description: "Notifications store unavailable."

  # ── API Tokens ─────────────────────────────────────────────────────────────
  /v1/api-tokens:
    get:
      operationId: listAPITokens
      summary: "List API tokens for the authenticated subject"
      tags:
        - api-tokens
      security:
        - sessionCookie: []
        - bearerToken: []
      responses:
        "200":
          description: "API tokens (hashed_token never returned)."
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/APIToken"
        "401":
          description: "Authentication required."
        "503":
          description: "API token store unavailable."
    post:
      operationId: createAPIToken
      summary: "Create a new API token"
      tags:
        - api-tokens
      security:
        - sessionCookie: []
        - bearerToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                scopes:
                  type: array
                  items:
                    type: string
              required:
                - name
      responses:
        "201":
          description: "Created token. The plainTextToken field is only returned in this response."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APITokenCreateResponse"
        "400":
          description: "Invalid input."
        "401":
          description: "Authentication required."
        "503":
          description: "API token store unavailable."
  /v1/api-tokens/{id}:
    delete:
      operationId: revokeAPIToken
      summary: "Revoke an API token by ID"
      tags:
        - api-tokens
      security:
        - sessionCookie: []
        - bearerToken: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: "Revoked token record."
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/APIToken"
        "400":
          description: "Invalid token ID."
        "401":
          description: "Authentication required."
        "404":
          description: "Token not found or does not belong to the caller."
        "503":
          description: "API token store unavailable."
...